+++ qt_utils.cpp
@@ -135,7 +135,7 @@ string convertToHTML(const string &str) {
 
 bool stringAnyNonWhitespace(const string &str) {
     bool any = false;
-    for(int i=0;i<str.length() && !any;i++) {
+    for(size_t i=0;i<str.length() && !any;i++) {
         char ch = str.at(i);
         if( ch != '\n' && ch != '\r' && !isspace(ch) ) {
             any = true;
